Separate GacelaConfigFileFactory into two classes
